cpu/intel/sandybridge: Add `hyper_threading` option
More and more people request an option to disable HT. To implement that
we have to toggle a bit in a `soft reset` register in the PCH that can
override certain default settings of the CPU when it comes out of reset.

The `soft reset` register is already used for other settings. So we have
to take care that all settings are gathered before we issue the reset.

Note, the current code using `soft reset` for flex ratio selection seems
incomplete.

Tested on ThinkPad X220 (i5-2520M) and ThinkPad W530 (i7-3940XM).

Change-Id: I2b73e32ff5af8ea64a47e8aa706e27648aaf0993
Signed-off-by: Nico Huber <nico.h@gmx.de>
Signed-off-by: Evgeny Zinoviev <me@ch1p.io>