drivers/apple: Add hybrid graphics driver
A hybrid graphics driver for Apple MacBook Pro.

The driver logic is based on lenovo/hybrid_graphics. It is splitted
into romstage and ramstage parts.

The mainboard code calls the driver from romstage to get the GPU
state. The driver reads the state from the `hybrid_grapihcs_mode`
nvram option, switches dGPU power on or off according to the state
and returns the state to the mainboard code. The mainboard code then
has to hide the disabled PCI device.

The ramstage part handles the graphics muxes. The muxes code is based
on the apple-gmux linux driver, originally written by:

* Canonical Ltd. <seth.forshee@canonical.com>
* Andreas Heider, 2010-2012 <andreas@meetr.de>
* Lukas Wunner, 2015 <lukas@wunner.de>

Tested on MacBook Pro Retina 15 Mid 2012 (MacBook Pro 10,1).

Change-Id: I22b66622cd2da0e9951ee726d650d204fbb8a5bc
Signed-off-by: Evgeny Zinoviev <me@ch1p.io>